[FEAT] 인코더 추가
